+++ app/src/main/java/com/battlelancer/seriesguide/lists/SgListFragment.kt
@@ -20,6 +20,8 @@ import com.battlelancer.seriesguide.databinding.FragmentListBinding
 import com.battlelancer.seriesguide.lists.ListsDistillationSettings.ListsSortOrderChangedEvent
 import com.battlelancer.seriesguide.lists.database.SgListItemWithDetails
 import com.battlelancer.seriesguide.provider.SeriesGuideContract.ListItemTypes
+import com.battlelancer.seriesguide.shows.episodes.EpisodeTools
+import com.battlelancer.seriesguide.shows.tools.ShowSync
 import com.battlelancer.seriesguide.ui.AutoGridLayoutManager
 import com.battlelancer.seriesguide.ui.OverviewActivity
 import com.battlelancer.seriesguide.ui.widgets.SgFastScroller
@@ -120,10 +122,14 @@ class SgListFragment : Fragment() {
             override fun onMenuClick(anchor: View, item: SgListItemWithDetails) {
                 val popupMenu = PopupMenu(anchor.context, anchor)
                 popupMenu.inflate(R.menu.lists_popup_menu)
+                val menu = popupMenu.menu
+                menu.findItem(R.id.menu_action_lists_favorites_add).isVisible = !item.favorite
+                menu.findItem(R.id.menu_action_lists_favorites_remove).isVisible = item.favorite
                 popupMenu.setOnMenuItemClickListener(
                     PopupMenuItemClickListener(
                         requireContext(), parentFragmentManager,
-                        item.listItemId, item.showId
+                        item.listItemId, item.showId,
+                        item.nextEpisode?.toLongOrNull() ?: 0
                     )
                 )
                 // Hide manage lists option for legacy show items, only allow removal.
@@ -144,18 +150,40 @@ class SgListFragment : Fragment() {
         private val context: Context,
         private val fragmentManager: FragmentManager,
         private val itemId: String,
-        private val showId: Long
+        private val showId: Long,
+        private val nextEpisodeId: Long
     ) : PopupMenu.OnMenuItemClickListener {
+
+        private val showTools = SgApp.getServicesComponent(context).showTools()
+
         override fun onMenuItemClick(item: MenuItem): Boolean {
-            val id = item.itemId
-            if (id == R.id.menu_action_lists_manage) {
-                ManageListsDialogFragment.show(fragmentManager, showId)
-                return true
-            } else if (id == R.id.menu_action_lists_remove) {
-                ListsTools.removeListItem(context, itemId)
-                return true
+            when (item.itemId) {
+                R.id.menu_action_lists_watched_next -> {
+                    EpisodeTools.episodeWatchedIfNotZero(context, nextEpisodeId)
+                    return true
+                }
+                R.id.menu_action_lists_favorites_add -> {
+                    showTools.storeIsFavorite(showId, true)
+                    return true
+                }
+                R.id.menu_action_lists_favorites_remove -> {
+                    showTools.storeIsFavorite(showId, false)
+                    return true
+                }
+                R.id.menu_action_lists_manage -> {
+                    ManageListsDialogFragment.show(fragmentManager, showId)
+                    return true
+                }
+                R.id.menu_action_lists_update -> {
+                    ShowSync.triggerDeltaSync(context, showId)
+                    return true
+                }
+                R.id.menu_action_lists_remove -> {
+                    ListsTools.removeListItem(context, itemId)
+                    return true
+                }
+                else -> return false
             }
-            return false
         }
     }